Changed file structure to make it easier to work with and to create easier tests.
