+++ generate-osm/db-utils/db-mapper.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"os"
 
+	"transform-osm/db-utils/mapper"
 	OSMUtil "transform-osm/osm-utils"
 
 	"github.com/pkg/errors"
@@ -28,7 +29,7 @@ func MapDB(
 	for _, line := range refs {
 		var anchors map[float64]([]*OSMUtil.Node) = map[float64]([]*OSMUtil.Node){}
 		var osm OSMUtil.Osm
-		var dbIss XmlIssDaten
+		var dbIss mapper.XmlIssDaten
 
 		osmLineFilePath := osmDir + "/" + line + ".xml"
 		osmFile, err := os.ReadFile(osmLineFilePath)
@@ -50,14 +51,14 @@ func MapDB(
 
 		fmt.Printf("Mapping line %s \n", line)
 
-		var notFoundSignalsFalling []*NamedSimpleElement = []*NamedSimpleElement{}
-		var notFoundSignalsRising []*NamedSimpleElement = []*NamedSimpleElement{}
-		var notFoundSwitches []*Weichenanfang = []*Weichenanfang{}
+		var notFoundSignalsFalling = []*mapper.NamedSimpleElement{}
+		var notFoundSignalsRising = []*mapper.NamedSimpleElement{}
+		var notFoundSwitches = []*mapper.Weichenanfang{}
 		var foundAnchorCount = 0
 		for _, stelle := range dbIss.Betriebsstellen {
 			for _, abschnitt := range stelle.Abschnitte {
 				for _, knoten := range abschnitt.Knoten {
-					err = findAndMapAnchorMainSignals(
+					err = mapper.FindAndMapAnchorMainSignals(
 						*knoten,
 						&osm,
 						anchors,
@@ -71,7 +72,7 @@ func MapDB(
 						return nil, nil, nil, -1, errors.Wrap(err, "failed anchoring main signals")
 					}
 
-					err = findAndMapAnchorSwitches(
+					err = mapper.FindAndMapAnchorSwitches(
 						*knoten,
 						&osm,
 						anchors,
@@ -94,10 +95,10 @@ func MapDB(
 		totalNumberOfAnchors += foundAnchorCount
 		totalElementsNotFound += numElementsNotFound
 
-		var issWithMappedSignals = XmlIssDaten{
-			Betriebsstellen: []*Spurplanbetriebsstelle{{
-				Abschnitte: []*Spurplanabschnitt{{
-					Knoten: []*Spurplanknoten{{
+		var issWithMappedSignals = mapper.XmlIssDaten{
+			Betriebsstellen: []*mapper.Spurplanbetriebsstelle{{
+				Abschnitte: []*mapper.Spurplanabschnitt{{
+					Knoten: []*mapper.Spurplanknoten{{
 						HauptsigF:  notFoundSignalsFalling,
 						HauptsigS:  notFoundSignalsRising,
 						WeichenAnf: notFoundSwitches,
@@ -118,7 +119,7 @@ func MapDB(
 			for _, stelle := range issWithMappedSignals.Betriebsstellen {
 				for _, abschnitt := range stelle.Abschnitte {
 					for _, knoten := range abschnitt.Knoten {
-						err = mapUnanchoredSignals(
+						err = mapper.MapUnanchoredSignals(
 							&osm,
 							anchors,
 							mainSignalList,
@@ -130,7 +131,7 @@ func MapDB(
 						if err != nil {
 							return nil, nil, nil, -1, errors.Wrap(err, "failed finding main signals")
 						}
-						err = mapUnanchoredSwitches(
+						err = mapper.MapUnanchoredSwitches(
 							&osm,
 							anchors,
 							&newNodeIdCounter,
@@ -144,8 +145,8 @@ func MapDB(
 				}
 			}
 
-			simpleElements := make(map[string]([]*SimpleElement))
-			namedSimpleElements := make(map[string]([]*NamedSimpleElement))
+			simpleElements := make(map[string]([]*mapper.SimpleElement))
+			namedSimpleElements := make(map[string]([]*mapper.NamedSimpleElement))
 
 			for _, stelle := range dbIss.Betriebsstellen {
 				for _, abschnitt := range stelle.Abschnitte {
@@ -159,7 +160,7 @@ func MapDB(
 						namedSimpleElements["track_end"] = knoten.Gleisende
 
 						for elementName, elementList := range simpleElements {
-							err = mapSimpleElement(
+							err = mapper.MapSimpleElement(
 								&osm,
 								anchors,
 								&newNodeIdCounter,
@@ -173,7 +174,7 @@ func MapDB(
 							}
 						}
 						for elementName, elementList := range namedSimpleElements {
-							err = mapNamedSimpleElement(
+							err = mapper.MapNamedSimpleElement(
 								&osm,
 								anchors,
 								&newNodeIdCounter,
@@ -187,7 +188,7 @@ func MapDB(
 							}
 						}
 						for _, signalType := range []string{"as", "ps"} {
-							err = mapUnanchoredSignals(
+							err = mapper.MapUnanchoredSignals(
 								&osm,
 								anchors,
 								otherSignalList,
@@ -201,7 +202,7 @@ func MapDB(
 							}
 						}
 
-						err = mapCrosses(
+						err = mapper.MapCrosses(
 							&osm,
 							anchors,
 							&newNodeIdCounter,
@@ -211,7 +212,7 @@ func MapDB(
 						if err != nil {
 							return nil, nil, nil, -1, errors.Wrap(err, "failed finding crosses")
 						}
-						err = mapHalts(
+						err = mapper.MapHalts(
 							&osm,
 							anchors,
 							haltList,
@@ -222,7 +223,7 @@ func MapDB(
 						if err != nil {
 							return nil, nil, nil, -1, errors.Wrap(err, "failed finding halts")
 						}
-						err = mapSpeedLimits(
+						err = mapper.MapSpeedLimits(
 							&osm,
 							anchors,
 							&newNodeIdCounter,
@@ -232,7 +233,7 @@ func MapDB(
 						if err != nil {
 							return nil, nil, nil, -1, errors.Wrap(err, "failed finding speed limits")
 						}
-						err = mapEoTDs(
+						err = mapper.MapEoTDs(
 							&osm,
 							anchors,
 							&newNodeIdCounter,
@@ -242,7 +243,7 @@ func MapDB(
 						if err != nil {
 							return nil, nil, nil, -1, errors.Wrap(err, "failed finding end of train detectors")
 						}
-						err = mapSlopes(
+						err = mapper.MapSlopes(
 							&osm,
 							anchors,
 							&newNodeIdCounter,
