+++ generate-osm/db-utils/mapper/map-switches_test.go
@@ -1,0 +1,15 @@
+package mapper_test
+
+import "testing"
+
+func TestFindAndMapAnchorSwitches(t *testing.T) {
+	// TODO
+}
+
+func TestMapUnanchoredSwitches(t *testing.T) {
+	// TODO
+}
+
+func TestMapCrosses(t *testing.T) {
+	// TODO
+}
