+++ generate-osm/db-utils/mapper/node-creation-export_test.go
@@ -1,0 +1,6 @@
+package mapper
+
+var CreateNamedDirectionalNode = createNamedDirectionalNode
+var CreateDirectionalNode = createDirectionalNode
+var CreateNamedSimpleNode = createNamedSimpleNode
+var CreateSimpleNode = createSimpleNode
