+++ generate-osm/osm-utils/generate-stations_test.go
@@ -11,68 +11,120 @@ func TestGenerateSearchFile(t *testing.T) {
 	type args struct {
 		osm osmUtils.Osm
 	}
-	test := struct {
+	type want struct {
+		stations map[string]osmUtils.Station
+		osm      osmUtils.Osm
+		err      error
+	}
+	tests := []struct {
 		name string
 		args args
+		want want
 	}{
-		name: "test the generation of the search file and adding the stations to the osm data",
-		args: args{
-			osmUtils.Osm{
-				Node: []*osmUtils.Node{
-					{
-						Id:  "1",
-						Lat: "1",
-						Lon: "2",
-						Tag: []*osmUtils.Tag{
-							{
-								K: "name",
-								V: "testStation",
-							},
-							{
-								K: "railway",
-								V: "station",
+		{
+			name: "test the generation of the search file and adding the stations to the osm data",
+			args: args{
+				osmUtils.Osm{
+					Node: []*osmUtils.Node{
+						{
+							Id:  "1",
+							Lat: "1",
+							Lon: "2",
+							Tag: []*osmUtils.Tag{
+								{
+									K: "name",
+									V: "testStation",
+								},
+								{
+									K: "railway",
+									V: "station",
+								},
 							},
 						},
-					},
-					{
-						Id:  "2",
-						Lat: "3",
-						Lon: "4",
-						Tag: []*osmUtils.Tag{
-							{
-								K: "name",
-								V: "testHalt",
-							},
-							{
-								K: "railway",
-								V: "halt",
+						{
+							Id:  "2",
+							Lat: "3",
+							Lon: "4",
+							Tag: []*osmUtils.Tag{
+								{
+									K: "name",
+									V: "testHalt",
+								},
+								{
+									K: "railway",
+									V: "halt",
+								},
 							},
 						},
 					},
 				},
 			},
+			want: want{
+				stations: map[string]osmUtils.Station{
+					"1": {
+						Name: "testStation",
+						Lat:  "1",
+						Lon:  "2",
+					},
+					"2": {
+						Name: "testHalt",
+						Lat:  "3",
+						Lon:  "4",
+					},
+				},
+				osm: osmUtils.Osm{
+					Node: []*osmUtils.Node{
+						{
+							Id:  "1",
+							Lat: "1",
+							Lon: "2",
+							Tag: []*osmUtils.Tag{
+								{
+									K: "name",
+									V: "testStation",
+								},
+								{
+									K: "railway",
+									V: "station",
+								},
+								{
+									K: "type",
+									V: "station",
+								},
+							},
+						},
+						{
+							Id:  "2",
+							Lat: "3",
+							Lon: "4",
+							Tag: []*osmUtils.Tag{
+								{
+									K: "name",
+									V: "testHalt",
+								},
+								{
+									K: "railway",
+									V: "halt",
+								},
+								{
+									K: "type",
+									V: "station",
+								},
+							},
+						},
+					},
+				},
+				err: nil,
+			},
 		},
 	}
 
-	t.Run(test.name, func(t *testing.T) {
-		stations, stationHaltsOsm := osmUtils.GenerateOsmAndSearchFile(test.args.osm)
-		assert.Equal(t, "testStation", stations["1"].Name, "Expected testStation, got %s", stations["1"].Name)
-		assert.Equal(t, "1", stations["1"].Lat, "Expected 1, got %s", stations["1"].Lat)
-		assert.Equal(t, "2", stations["1"].Lon, "Expected 2, got %s", stations["1"].Lon)
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			stations, stationHaltsOsm := osmUtils.GenerateOsmAndSearchFile(test.args.osm)
 
-		assert.Equal(t, "testHalt", stations["2"].Name, "Expected testHalt, got %s", stations["2"].Name)
-		assert.Equal(t, "3", stations["2"].Lat, "Expected 3, got %s", stations["2"].Lat)
-		assert.Equal(t, "4", stations["2"].Lon, "Expected 4, got %s", stations["2"].Lon)
-
-		for _, stationHaltNode := range stationHaltsOsm.Node {
-			if stationHaltNode.Id == "1" {
-				assert.Equal(t, "type", stationHaltNode.Tag[2].K, "Expected type, got %s", stationHaltNode.Tag[2].K)
-				assert.Equal(t, "station", stationHaltNode.Tag[2].V, "Expected station, got %s", stationHaltNode.Tag[2].V)
-			}
-			if stationHaltNode.Id == "2" {
-				assert.Equal(t, "type", stationHaltNode.Tag[2].K, "Expected type, got %s", stationHaltNode.Tag[2].K)
-				assert.Equal(t, "station", stationHaltNode.Tag[2].V, "Expected element, got %s", stationHaltNode.Tag[2].V)
-			}
-		}
-	})
+			assert.Equal(t, test.want.stations, stations, "Expected %v, got %v", test.want.stations, stations)
+			assert.EqualValues(t, test.want.osm, stationHaltsOsm, "Expected %v, got %v", test.want.osm, stationHaltsOsm)
+		})
+	}
 }
